refactor(runtime): isolate package transaction ownership
